[MCJ-1600] refactor: 관리자 공구 요청 패키지를 admin 도메인으로 정리
