+++ src/main/kotlin/com/moongchijang/domain/admin/application/dto/AdminOwnerGroupBuyCloseRequestRejectRequest.kt
@@ -1,4 +1,4 @@
-package com.moongchijang.domain.owner.application.dto
+package com.moongchijang.domain.admin.application.dto
 
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
